requirement.txt
pdfplumber
openai
python-docx

